requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
amqp==5.1.1
APScheduler==3.8.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.13.0
django-heroku==0.3.1
django-simple-history==3.0.0
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.3
frozenlist==1.3.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.0
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.2
packaging==21.3
prompt-toolkit==3.0.31
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
yarl==1.7.2
zope.interface==5.4.0
